[add]438 Viewing a Single Post

+++ utility/remove-images.js
@@ -1,6 +1,9 @@
 const path = require("path");
 const fs = require("fs");
 module.exports = (filePath) => {
+  if (!filePath) {
+    throw new Error("failded to Delete the image. currupted path.");
+  }
   filePath = path.join(__dirname, "..", filePath);
   fs.unlink(filePath, (err) => {
     if (err) {
